Move AST and fix binary operator parser
Problem: Our binary operator parser always builds a right-balanced tree.
Oops.

Solution: Introduce ambiguities so that a shift or reduce can take place
to properly balance the tree with the `%left`, `%nonassoc`, and `%right`
directives.

Make `in`, `of`, `do`, `else` and `:=` extend as far to the right as
possible to remove all shift/reduce conflicts.

Move AST to its own module to be more organized.

Technically, we could get rid of all expression levels and put
everything in `Exp`, but like this we have a few less states and it
looks more organized.

Remove redudant `$$`s from the token patterns.