Add a typechecker
Problem: This should be most of the part "a" of the typechecker, which
doesn't yet handle recursive datatypes or check if break is contained
within loops.

Solution: Use most of what is discussed in the books, as well as my
intuition to make things work. Maybe it's not working as well yet, but
at least one of the programs typechecks correctly. The other still
requires part b to be done.